Create SQL_PRUEBA.spsql
# Junior Software Engineering (data) test

[treble.ai](http://treble.ai) uses a Redshift datawarehouse that behaves very similar to a postgreSQL database. In this exercise assume you are working with standard Postgresql

Once you are connected you will see the following tables.

### Company

```
Table "public.company"
       Column       |            Type             | Collation | Nullable | Default | Storage  | Stats target | Description 
--------------------+-----------------------------+-----------+----------+---------+----------+--------------+-------------
 id                 | integer                     |           | not null |         | plain    |              | 
 name               | character varying(255)      |           |          |         | extended |              | 
 associated_partner | character varying(255)      |           |          |         | extended |              | 
 close_date         | timestamp without time zone |           |          |         | plain    |              |
```

Associated parter tells if the company was brought by the partnerships team, and it has the value of the name of the associated partner if it was brought by the team, otherwise it is empty.

The close_date property tells the day the company signed a contract with treble.

### Conversations

```
Table "public.conversations"
   Column   |  Type   | Collation | Nullable | Default | Storage | Stats target | Description 
------------+---------+-----------+----------+---------+---------+--------------+-------------
 account_id | integer |           | not null |         | plain   |              | 
 total      | integer |           | not null |         | plain   |              | 
 date       | date    |           | not null |         | plain   |              | 
 successful | boolean |           | not null |         | plain   |              | 
 id         | integer |           |          |         | plain   |              |
```

A company may have multiple treble accounts, in each account the company can deploy conversations. In the table conversations it groups the conversations by day, account and “successful” category. A set of conversations can be successful or not, the categorization has already been done, so each row tells how many conversations per day and account were successful and how many were not successful.

### Stripe Invoce

```
Table "public.stripe_invoice"
   Column   |            Type             | Collation | Nullable | Default | Storage  | Stats target | Description 
------------+-----------------------------+-----------+----------+---------+----------+--------------+-------------
 sent_date  | timestamp without time zone |           | not null |         | plain    |              | 
 invoice_id | character varying(255)      |           | not null |         | extended |              | 
 company_id | character varying(255)      |           | not null |         | extended |              | 
 amount     | integer                     |           | not null |         | plain    |              | 
 currency   | character varying(50)       |           | not null |         | extended |              |
```

Stripe is the system treble uses to send invoices to the companies. Each company may have multiple stripe accounts. The company id column in this table is the internal id stripe creates per each stripe account.

### Company Identifiers

```
Table "public.company_identifiers"
       Column       |          Type          | Collation | Nullable | Default | Storage  | Stats target | Description 
--------------------+------------------------+-----------+----------+---------+----------+--------------+-------------
 account_identifier | integer                |           | not null |         | plain    |              | 
 company_id         | integer                |           | not null |         | plain    |              | 
 stripe_company_ids | character varying(255) |           |          |         | extended |              |
```

This table have the association between the conversations, stripe and company tables. It tells per company which are the accounts ids from the treble accounts and stripe accounts.

# Exercise

The partnerships team, has the goal of activating companies. 

Activation occurs on the day a company, aggregating across all its accounts, initiates 350 or more conversations within any given three-day period. For instance, consider a company that launches its operations and records the following daily conversation counts from October 1st to October 7th: 100, 100, 100, 100, 100, 150, 100. In this scenario, the activation day would be recognized on October 6th, as it marks the culmination of a three-day window (October 4th, 5th, and 6th) when the conversation tally first meets or surpasses the 350 threshold. Any conversation counts, regardless if the conversations were successful or not.

Be aware that there could be days when the company has no records on the table conversations. So for example if the company deploys from October 7th to October 10th: 50, 100, 0, 200. There won’t be any records for October 9th, data will look as follows

```
 account_id | total |    date    | successful |  id  
------------+-------+------------+------------+------
          1 |    50 | 2023-10-07 | t          | 3663
          1 |    90 | 2023-10-08 | t          | 3659
          1 |    10 | 2023-10-08 | f          | 3660
          1 |   150 | 2023-10-10 | t          | 3656
          1 |    50 | 2023-10-10 | f          | 3652
```

Your task is to build a query that retrieves per company_id the activation date. you **cannot** insert tables into the database, and you cannot use the **generate_series** function. The use of control-flow structures commonly used in procedural programming languages (like "while" loops, "for" loops, etc.) is NOT permitted. You are expected to solve these problems using standard SQL statements and capabilities. Solutions employing non-SQL constructs will not be accepted.

# Deliverable

For this task, send the solution (query) to the corporative email.